requirements.txt
pandas
duckdb
unidecode
psycopg2-binary
